requirements.txt
numpy==1.11.0
pyalsaaudio==0.8.2
redis==2.10.5